Move database management to separate folder from models
